[stable/traefik] Use kvprovider prefix in storage path
The `storage` configuration was previously hardcoded. When using
multiple traefik installations, sharing the same kvprovider backend
they would all write their certificates to the same location. This
change makes the `storage` configuration conform to the "prefix"
configured for the kvprovider.

I have also added the fullname to the storeconfig-job, to ensure
that multiple instances of traefik doesn't clash.

Signed-off-by: Frederik Nordahl Jul Sabroe <frederikns@gmail.com>